requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
python-docx>=1.1
openai>=1.30
anthropic>=0.30
Pillow>=10.0
python-dotenv>=1.0
